moved majority of bodytesthelper commands
