+++ cypress/integration/articleBodySpec.js
@@ -1,12 +1,4 @@
 import config from '../support/config/services';
-import {
-  copyrightDataWindow,
-  firstHeadlineDataWindow,
-  firstParagraphDataWindow,
-  firstSubheadlineDataWindow,
-  placeholderImageLoaded,
-  renderedTitle,
-} from '../support/bodyTestHelper';
 
 const serviceHasArticlePageType = service =>
   config[service].pageTypes.articles !== undefined;
@@ -23,7 +15,7 @@ Object.keys(config)
       });
 
       it('should render an H1, which contains/displays a styled headline', () => {
-        firstHeadlineDataWindow();
+        cy.firstHeadlineDataWindow();
       });
 
       it('should render a formatted timestamp', () => {
@@ -41,13 +33,13 @@ Object.keys(config)
       it('should render an H2, which contains/displays a styled subheading', () => {
         cy.window().then(win => {
           if (win.SIMORGH_DATA.pageData.metadata.language === 'en-gb') {
-            firstSubheadlineDataWindow();
+            cy.firstSubheadlineDataWindow();
           }
         });
       });
 
       it('should render a paragraph, which contains/displays styled text', () => {
-        firstParagraphDataWindow();
+        cy.firstParagraphDataWindow();
       });
 
       it('should have a placeholder image', () => {
@@ -92,16 +84,16 @@ Object.keys(config)
       });
 
       it('should have an image copyright label with styling', () => {
-        copyrightDataWindow();
+        cy.copyrightDataWindow();
       });
 
       it('should render a title', () => {
         cy.window().then(win => {
           const { seoHeadline } = win.SIMORGH_DATA.pageData.promo.headlines;
           if (win.SIMORGH_DATA.pageData.metadata.language === 'en-gb') {
-            renderedTitle(`${seoHeadline} - BBC News`);
+            cy.renderedTitle(`${seoHeadline} - BBC News`);
           } else {
-            renderedTitle(`${seoHeadline} - BBC News فارسی`);
+            cy.renderedTitle(`${seoHeadline} - BBC News فارسی`);
           }
         });
       });
