+++ cypress/integration/metaArticleSpec.js
@@ -1,7 +1,6 @@
 import envConfig from '../support/config/envs';
 import services from '../support/config/services';
 import testData from '../../src/app/lib/config/services';
-import { hasHtmlLangDirAttributes } from '../support/bodyTestHelper';
 import {
   checkAmpHTML,
   checkCanonicalURL,
@@ -28,7 +27,7 @@ Object.keys(services)
       });
 
       it('should have the correct lang & dir attributes', () => {
-        hasHtmlLangDirAttributes({
+        cy.hasHtmlLangDirAttributes({
           lang: `${testData[service].datetimeLocale}`,
           dir: `${testData[service].dir}`,
         });
