+++ cypress/support/commands.js
@@ -1,3 +1,5 @@
+import './commands/pageTestFragment/articleBody';
+
 // Overwriting Cypress Commands should very rarely be done.
 Cypress.Commands.overwrite('visit', (originalFn, url, options) => {
   cy.request({ url, failOnStatusCode: false }).then(({ headers }) => {
